Fix bug qd phe duyet kq bdg, bb bdg

+++ src/main/java/com/tcdt/qlnvhang/request/bandaugia/bienbanbandaugia/BhBbBanDauGiaReq.java
@@ -11,7 +11,7 @@ import java.util.List;
 @NoArgsConstructor
 public class BhBbBanDauGiaReq {
     private Long id;
-    private Long nam;
+    private Integer nam;
     private String soBienBan;
     private String trichYeu;
     private LocalDate ngayKy;
